--- src/main/java/com/team19/musuimsa/exception/auth/UserAccessDeniedException.java
@@ -1,8 +1,0 @@
-package com.team19.musuimsa.exception.auth;
-
-public class UserAccessDeniedException extends AuthenticationException {
-
-    public UserAccessDeniedException(String message) {
-        super(message);
-    }
-}
